% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeAll.R
\name{analyzeAll}
\alias{analyzeAll}
\title{analyzeAll}
\usage{
analyzeAll(
  experiment = NULL,
  parameters = NULL,
  save = TRUE,
  analyze = c("ASVs", "OTUs", "KOs", "pathways"),
  dissimilarities = c("0.03"),
  analyzes = c("alpha-diversity", "beta-diversity", "taxonomy", "diff-features"),
  basename = "bacteria",
  outdir = ".",
  graphics.type = "ggplot",
  vertical = TRUE,
  collect = FALSE,
  tag_level = "A",
  statistics = TRUE,
  rand.seed = 667,
  permu = 999,
  significance.threshold = 0.05,
  find.differing.pairs = TRUE,
  palette.family = "ggsci",
  palette = "nrc_npg",
  processors = 1,
  width = 3.5,
  height = 3.5,
  main.variable = NULL,
  facetting.variables = NULL,
  varpart.variable = NULL,
  sample.color.variable = NULL,
  sample.size.variable = NULL,
  feature.color.variable = NULL,
  chemistry.variables = NULL,
  alpha.div.measures = c("sobs", "H", "E"),
  alpha.div.points = T,
  ref.group = NULL,
  alpha.plot.style = "box",
  alpha.plot.cols = 2,
  beta.diversity.metrics = c("bray", "GUniFrac"),
  unifrac.types = c("d_0.5"),
  beta.diversity.analyzes = c("nmds", "dbrda"),
  biplot = T,
  triplot = F,
  size.means.abundance = T,
  abundance.cutoff = 50,
  ellipses = T,
  vector.col = "red",
  ordistep.steps = 500,
  ordistep.direction = "both",
  diff.features.methods = c("deseq"),
  log2FC.threshold = 1,
  mergeby = 0,
  consensus = F,
  plot.diff.features = F,
  num.top.features = 10,
  write.single = T,
  write.final = T,
  mc.samples = 128,
  rare.threshold = 0.01,
  taxlevel.names = c("kingdoms", "phyla", "classes", "orders", "families", "genera")
)
}
\arguments{
\item{experiment}{An mg-class experiment.}

\item{parameters}{A list or a path to a tsv file with parameters. Any other function parameter may apart from experiment may be set in this object/file. If parameters are set both in file/object, the settings can be overriden in the function call. Defaults to NULL.}

\item{analyze}{What to analyze? A vector of strings being names of features to be analyzed. Current possibilities: 'ASVs', 'KOs', 'pathways', 'OTUs'. Defaults to c('ASVs', "OTUs', 'KOs', 'pathways').}

\item{dissimilarities}{A vector of numbers between 0 and 1 determining for which dissimilarities OTUs should be analyzed (and generated, if necessary). Defaults to c(0.03).}

\item{analyzes}{How to analyze? A vector of strings being names of analyzes to be performed. Current possibilities: 'alpha-diversity', 'beta-diversity', 'taxonomy', 'diff-features'. Defaults to c('alpha-diversity', 'beta-diversity', 'taxonomy', 'diff-features')}

\item{basename}{What should be a common part of produced files' names? Defaults to 'bacteria'.}

\item{outdir}{A path to directory where resulting plots and other files should be stored. Defaults to ".".}

\item{graphics.type}{Which graphics type should be used? One of 'ggplot' or 'base'. Defaults to 'ggplot'.}

\item{vertical}{A logical determining if compound plots should be organized vertically (i.e. their width be smaller than height). Defaults to TRUE.}

\item{collect}{A logical determining if identical legends should be collected in one place of a compound plot. Defaults to FALSE.}

\item{statistics}{Should statistical tests be performed? A logical. Defaults to TRUE.}

\item{rand.seed}{A random seed for reproducibility. Must be an odd positive integer. Defaults to 667 (a new, better beast).}

\item{permu}{A positive integer determining the number of permutations performed during non-parametric tests such as PERMANOVA. Defaults to 999.}

\item{significance.threshold}{A number between 0 and 1 determining significance threshold for statistical tests. Defaults to 0.05.}

\item{find.differing.pairs}{Should pairwise tests be performed to find differing group pairs if overall test is significant? A logical. Defaults to TRUE.}

\item{palette.family}{Which palette family should be used? Any family included in the paletteer package may be used. Defaults to 'ggsci'.}

\item{palette}{Which palette should be used? It should be a name of a palette from family chosen with palette.family. Any palette from the paletteer package may be used here.  Defaults to 'npg-nrc' from the 'ggsci' family.}

\item{processors}{Number of processors to use. Defaults to 1.}

\item{width}{A number determining width of a single plot (in inches). Defaults to 3.5.}

\item{height}{A number determining height of a single plot (in inches). Defaults to 3.5.}

\item{main.variable}{The main variable to be analyzed. A string indicating a categorical variable from experiment's sampledata. Absolutely necessary. It will be mapped to symbol shape on beta-diversity plots and used as grouping variable in all other analyzes. Defaults to NULL.}

\item{facetting.variables}{Should facetted composite plots be produced? A vector of one or two variable names. Optional. The names must be chosen from colnames(sampledata) or variableNames(experiment). Defaults to NULL (no facetting).}

\item{varpart.variable}{A string indicating which variable from sampledata should be used in variance partitioning to enable assessing main variable's influence. Defaults to NULL.}

\item{sample.color.variable}{A string indicating which variable from sampledata should be mapped to color of symbols denoting samples. Optional. Must be categorical. Defaults to NULL.}

\item{sample.size.variable}{A string indicating which variable from sampledata should be mapped to size of symbols denoting samples. Optional. Must be continuous. Defaults to NULL.}

\item{feature.color.variable}{A string indicating which variable from featureannot should be mapped to color of symbols representing features. Must be categorical}

\item{alpha.div.measures}{Which species richness, diversity and evenness metrics should be used in alpha-diversity analysis? A vector of index names. Current possibilities: i) species richness: 'sobs' - stands for 'species observed'; 'chao' - Chao1 index (estimated total richness); 'ace' - ACE index (estimated total richness); ii) diversity: 'shannon' - Shannon's H'; 'simpson' - Simpson's D; 'invsimpson' - Simpson's 1/D; iii) evenness: 'shannoneven' - Shannon's E. Defaults to c('sobs', 'shannon', 'shannoneven').}

\item{alpha.div.points}{A logical determining if poits representing samples should be added to alpha diversity plots. Defaults to TRUE.}

\item{alpha.plot.style}{A string determining how alpha diversity will be plotted. Current possibilities: 'box', 'violin', 'dot'. Defauts to 'box'.}

\item{alpha.plot.cols}{A positive integer determining number of columns in compound alpha diversity plots. Defaults to 2.}

\item{beta.diversity.analyzes}{Which beta-diversity analyzes should be performed? Current possibilities: 'nmds', 'rda', 'dbrda'. Defaults to c('nmds', 'dbrda')}

\item{biplot}{A logical determining if biplots should be generated. Defaults to TRUE.}

\item{triplot}{A logical determining if triplots should be generated. Requires giving at least one variable in chemical.variables. Defaults to FALSE.}

\item{size.means.abundance}{Should size of feature symbols on a biplot convey features' abundance? Defaults to TRUE.}

\item{abundance.cutoff}{How many features should be displayed on a biplot? Most abundant features are displayed. An integer greater than 1. Defaults to 50.}

\item{diff.features.methods}{Which methods should be used to find differentially abundant features? A vector of method names. Current possibilities: 'deseq' (DESeq2 is used). Defaults to c('deseq').}

\item{log2FC.threshold}{A number determining a minimal absolute value of log2 fold change of a feature required to consider the feature differentially abundant. Defaults to 1 (meaning two-fold difference).}

\item{rare.threshold}{Threshold for identification of rare features. All features whose abundance is less than threshold will be lumped to a category named 'rare'. Must be a number between 0 and 1. Defaults to 0.01.}

\item{chemical.variables}{A vector of strings indicating which variables from sampledata should be used in CCA or (db)RDA. Defaults to NULL.}

\item{beta.metrics}{Which beta-diversity metrics should be used? GUniFrac or any distance possible to calculate with 'vegdist' can be used. Defaults to c('bray', 'GUniFrac') - Bray-Curtis and GUniFrac.}

\item{unifracs}{Which variants of Generalized UniFrac should be used as beta-diversity metrics? ... Defaults to c('d_05').}
}
\value{
An mg-class experiment object with 'analyzes' slots populated with results of chosen analyzes.
}
\description{
Analyze all features in an mg-class experiment.
}
\details{
The function performs analyzes on various types of features resulting from one experiment. Currently the set of possible features comprises ASVs, OTUs, KO functions and pathways (from PICRUSt2 analysis). In the future other types of features will be added, such as genes or transcripts.
}
