% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeFeature.R
\name{analyzeFeature}
\alias{analyzeFeature}
\title{analyzeFeature}
\usage{
analyzeFeature(
  experiment = NULL,
  featurename = NULL,
  d = c(0.03),
  featureSet = NULL,
  featuretab = NULL,
  sampledata = NULL,
  featureannot = NULL,
  tree = NULL,
  basename = "",
  directory = ".",
  rarefied = NULL,
  depth = "auto",
  analyzes = c("alpha-diversity", "beta-diversity", "taxonomy", "diff-features"),
  graphics.type = "ggplot",
  statistics = T,
  permu = 999,
  significance.threshold = 0.05,
  find.differing.pairs = T,
  rand.seed = 667,
  variable = NULL,
  facetting.variables = NULL,
  vertical = T,
  collect = F,
  comps = NULL,
  processors = 4,
  palette.family = "ggsci",
  palette = "nrc_npg",
  tag_level = NULL,
  width = 3.5,
  height = 3.5,
  alpha.div.measures = c("sobs", "H", "E"),
  alpha.plot.style = "box",
  reflevel = NULL,
  alpha.plot.cols = 2,
  alpha.subplots.annotation = NULL,
  points = F,
  beta.diversity.analyzes = c("nmds", "dbrda"),
  beta.diversity.metrics = c("bray", "horn", "GUniFrac"),
  unifrac.types = c("d_0.5"),
  sample.color.variable = NULL,
  sample.size.variable = NULL,
  ellipses = T,
  biplot = T,
  feature.color.variable = NULL,
  size.means.abundance = T,
  abundance.cutoff = 50,
  triplot = F,
  chemistry.variables = NULL,
  vector.col = "red",
  ordistep.steps = 1000,
  ordistep.direction = "both",
  varpart.variable = NULL,
  log2FC.threshold = 1,
  mergeby = 0,
  diff.abund.methods = c("deseq"),
  diff.abund.taxlevels = c("phyla", "classes", "orders", "families", "genera"),
  consensus = F,
  plot.diff.features = F,
  num.top.features = 10,
  write.single = F,
  write.final = T,
  mc.samples = 128,
  taxlevel.names = c("phyla", "classes", "orders", "families", "genera"),
  rare.threshold = 0.01
)
}
\arguments{
\item{experiment}{An mg-class experiment object.}

\item{featurename}{A string indicating a name of a feature to analyze. Currently one of 'ASVs', 'OTUs', 'KOs', or 'pathways'.}

\item{d}{A vector of OTU dissimilarities. If featurename is 'OTUs' the argument determines which dissimilarities are analyzed. Defaults to c('0.03')}

\item{featureSet}{A list consisting of four elements: featuretab (matrix), sampledata (data frame), featureannot (data frame), and tree (phylo). A featureset can be extracted from an mg-class experiment with extractFeatureSet(). Disregarded if an experiment object is given.}

\item{featuretab}{A named integer matrix with samples in rows and features (ASVs, OTUs, genes, etc.) in columns. Can be extracted from an mg-class object with featuretab(). Disregarded if an experiment or featureset is given.}

\item{sampledata}{A data frame with samples in rows and environmental and grouping variables in columns. Can be extracted from an mg-class object with sampledata(). Disregarded if an experiment or featureset is given.}

\item{featureannot}{A data frame with features in rows and annotating variables (usually taxonomy or function) in columns.Can be extracted from an mg-class object with featureannot(). Disregarded if an experiment or featureset is given.}

\item{tree}{A phylo-class object or a path to a nexus file containing a tree. Can be extracted from an mg-class object with tree(). Disregarded if an mg-class object or featureset is given.}

\item{basename}{A string with which all filenames produced by the function will begin. Defaults to "".}

\item{directory}{A string being a valid path to a directory where resulting files will be stored. Defaults to ".", which means current directory.}

\item{rarefied}{If a featuretab/featureset is rarefied it should be set to TRUE to indicate this fact. Disregarded if an mg-class object is given. Defaults to NULL.}

\item{depth}{'auto' or an integer indicating to what depth data should be rarefied. If 'auto' depth is calculated to maximize information content in rarefied data (see rarefyData()). Disregarded if an mg-class object is given or when featuretab/featureset is rarefied.}

\item{analyzes}{A vector of analyzes' names. Current possibilities: 'alpha-diversity', 'beta-diversity', 'taxonomy', 'diff-features'. Defaults to c('alpha-diversity', 'beta-diversity', 'taxonomy', 'diff-features').}

\item{graphics.type}{A string indicating if ggplot2 or base graphics should be used. Current possibilities: 'ggplot', 'base'. Defaults to 'ggplot'.}

\item{statistics}{A logical indicating if statistical tests should be performed. Defaults to TRUE.}

\item{permu}{An integer indicating a number of permutations used in statistical tests. Defaults to 999.}

\item{significance.threshold}{A significance threshold for tests. Defaults to 0.05.}

\item{find.differing.pairs}{A logical indicating if pairwise tests should be performed in case there are more than two groups and overall p-value is below significance.threshold. Defaults to TRUE.}

\item{rand.seed}{An odd integer being a seed for a random numbers generator. Defaults to 667 (a new better beast).}

\item{variable}{A string indicating name of a main variable to be analyzed. It must be one of categorical variables in sampledata. The variable is mapped to shape of symbols representing samples on beta diversity plots. Defaults to '' - must be set.}

\item{facetting.variables}{A vector of strings indicating maximum two variables according to which data should be divided (and displayed on plot panels). Defaults to NULL.}

\item{vertical}{A logical indicating if facetted plots should be organized vertically - i.e. height should be greater than width -  (TRUE) or horizontally (FALSE). Defaults to TRUE.}

\item{collect}{A logical indicating if common legends should be collected in one place. Defaults to FALSE.}

\item{comps}{A vector of strings indicating which groups should be compared. Group names must be levels of the main variable. Defaults to NULL (all possible comparisons).}

\item{processors}{An integer indicating number of processors to use. Defaults to 1.}

\item{palette.family}{A string indicating which family of palettes should be used. Any family included in the paletteer packages can be used. Defaults to 'ggsci'.}

\item{palette}{A string indicating a palette from palette.family. Defaults to 'nrc-npg'.}

\item{tag_level}{A string indicating if and how subplots of a plot should be marked. Current possibilities: NULL - no tagging, 'A' - capital letters, 'a' - small letters , '1' - arabic numbers, 'I' - capitalized Roman numbers, 'i' - small Roman numbers. Defaults to NULL.}

\item{width}{A number indicating width of an individual plot (in inches). Defaults to 3.5.}

\item{height}{A number indicating height of an individual plot (in inches). Defaults to 3.5.}

\item{alpha.div.measures}{A vector of strings indicating which alpha diversity measures should be used. Current possibilities: i) species richness: 'sobs' - observed richness, 'chao' - estimated total richness (Chao1 index), 'ace' - estimated total richness (ACE index); ii) diversity: 'H' - Shannon's H', 'simpson' - Simpson's D, 'invsimpson' - Simpson's 1/D; iii) evennes: 'E' - Shannon's E. Defaults to c('sobs', 'H', 'E').}

\item{alpha.plot.style}{A string indicating how alpha diversity should be plotted. Current possibilities: 'box', 'violin', 'dot'. Defaults to 'box'.}

\item{reflevel}{NULL or a string indicating a reference group for alpha diversity analysis to compare all others with. Defaults to NULL.}

\item{alpha.plot.cols}{An integer indicating how many columns there should be in alpha diversity plots. Defaults to 2.}

\item{alpha.subplots.annotation}{A string indicating if and how individual subplots of an alpha diversity plot should be tagged. Current possibilities: NULL - no tagging, 'A' - capital letters, 'a' - small letters , '1' - arabic numbers, 'I' - capitalized Roman numbers, 'i' - small Roman numbers. Defaults to NULL.}

\item{points}{A logical indicating if points representing individual samples should be added to alpha diversity box- or violin plots. Defaults to TRUE.}

\item{beta.diversity.analyzes}{A vector of strings indicating which beta diversity analyzes should be performed. Current possibilities: 'nmds', 'rda', 'dbrda', 'cca'. Defaults to c('nmds', 'dbrda')}

\item{beta.diversity.metrics}{A vector of strings indicating which beta diversity metrics (indices) should be used. Any metric which can be calculated with vegan's vegdist or GUniFrac. Defaults to c('bray', 'GUniFrac').}

\item{unifrac.types}{A vector of strings indicating which flavors of GUniFrac should be used. Current possibilities: 'd_x' where x is a float, 'd_UW' weighted UniFrac. Defaults to c('d_0.5').}

\item{sample.color.variable}{A string indicating which variable should be mapped to color of symbols representing samples. It must be one of categorical variables in sampledata. Defaults to ''.}

\item{sample.size.variable}{A string indicating which variable should be mapped to size of symbols representing samples. It must be one of continuous variables in sampledata. Defaults to ''.}

\item{ellipses}{A logical indicating if 95\% CI ellipses should be drawn on beta diversity plots. Defaults to TRUE.}

\item{biplot}{A logical indicating if biplots (plots with samples and feature) should be produced in beta diversity analyses. Defaults to TRUE.}

\item{feature.color.variable}{A string indicating which variable from the featureannot table should be mapped to color of symbols representing features. Defaults to ''.}

\item{size.means.abundance}{A logical indicating if feature abundance should be mapped to symbol size on beta diversity plots. Defaults to TRUE.}

\item{abundance.cutoff}{A number indicating how many features with greatest abundance should appear on a biplot. Defaults to 10.}

\item{triplot}{A logical indicating if a triplot should be produced in beta diversity analyses. If TRUE at least one of 'rda', 'dbrda', 'cca' needs to be included in beta.diversity.analyzes.}

\item{chemistry.variables}{NULL or a vector of strings indicating which variables from sampledata should be used as covariates in beta.diversity analyses. Defaults to NULL.}

\item{vector.col}{A string indicating color in which vectors representing significant environmental covariates should be plotted. Defaults to 'red'.}

\item{ordistep.steps}{An integer indicating how many steps of model selection should be performed. Defaults to 1000.}

\item{ordistep.direction}{A string indicating algorithm for model selection. Current possibilities: 'forward', 'backward', 'both'. Defaults to 'both'.}

\item{varpart.variable}{NULL or a string indicating which variable should be used in variance partitioning (to allow computing main variable's explained variance). Defaults to NULL (no variance partitioning performed).}

\item{log2FC.threshold}{A number indicating minimal value of abs(log2 fold change) needed to consider a feature differentially abundant. Defaults to 1 (two-fold difference).}

\item{mergeby}{An integer indicating by which column should results of differential analysis be joined with features annotation in featureannot. Defaults to 0 (row (i.e. feature) names).}

\item{diff.abund.methods}{A vector of strings with names of methods used for differential abundance analysis. Current possibilities: 'deseq' - DESeq2, 'aldex' - ALDEx2. Defaults to c('deseq').}

\item{consensus}{A logical indicating if only features found to be differentially abundant by all chosen methods should be regarded significant. Defaults to FALSE.}

\item{plot.diff.features}{A logical indicating if differentially abundant features should be plotted. Defaults to FALSE.}

\item{num.top.features}{An integer indicating how many significant differentially abundant features to plot. Defaults to 10.}

\item{write.single}{A logical indicating if tables with differentially abundant features should be written out for each comparison (pair of groups). Defaults to TRUE.}

\item{write.final}{A logical indicating if final table with differentially abundant features (containing all comparisons) should be written to a file. Defaults to TRUE.}

\item{mc.samples}{A number of MCMC samples in ALDEx2. Defaults to 128.}

\item{taxlevel.names}{A vector of strings with taxonomical level names. Defaults to c('phyla', 'classes', 'orders', 'families', 'genera').}

\item{rare.threshold}{A number between 0 and 1 indicating maximal abundance of a feature categorized as 'rare'. Defaults to 0.01}
}
\value{
An mg-class experiment object(s) with 'analyzes' slot for a given feature populated with results of analyzes.
}
\description{
Analyze single feature from an mg-class experiment object.
}
