% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeAll.R
\name{check_and_generate_data}
\alias{check_and_generate_data}
\title{check_end_generate_data}
\usage{
check_and_generate_data(
  experiment = NULL,
  analyze = NULL,
  analyzes = NULL,
  dissimilarities = NULL,
  processors = 1
)
}
\arguments{
\item{experiment}{An mg-class object}

\item{analyze}{A vector of strings specifying which features should be analyzed}

\item{analyzes}{A vector of strings specifying which analyzes should be performed}

\item{dissimilarities}{A vector of strings specifying at which dissimilarities OTUs should be constructed}
}
\value{
An mg-class objects with OTUs/PICRUSt2 results
}
\description{
check_end_generate_data
}
