% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{cleanData}
\usage{
cleanData(
  featuretab = NULL,
  featureannot = NULL,
  remove.rare = NULL,
  remove.taxa = NULL,
  keep.only = NULL
)
}
\arguments{
\item{featuretab}{A named integer matrix with samples in rows and features (ASVs, OTUs, genes, KO functions etc.) in columns. Can be extracted from an mg-class object with featuretab(). Defaults to NULL.}

\item{featureannot}{A data frame with features in rows and annotations (usually taxonomy or functions) in columns. Can be extracted from an mg-class object with featureannot(). Defaults to NULL.}

\item{remove.rare}{An integer indicating maximal total abundance of features removed as rare (thus probably artefactual). Defaults to NULL.}

\item{remove.taxa}{A vector of strings indicating which taxa are contaminants and should be removed. Defaults to NULL.}

\item{keep.only}{A string indicating a taxon to keep. Defaults to NULL.}
}
\value{
List with two elements: featuretab and featureannot.
}
\description{
Remove artifacts (rare variants) and unwanted taxa.
}
\examples{
NULL
}
