% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_plot.R
\name{compose_plot}
\alias{compose_plot}
\title{compose_plot}
\usage{
compose_plot(
  object = NULL,
  sampledata = NULL,
  featurename = NULL,
  file.name = "",
  d = NULL,
  analysis = NULL,
  variables = NULL,
  vertical = NULL,
  collect = F,
  tag_level = "A",
  width = 3.5,
  height = 3.5
)
}
\arguments{
\item{object}{An mg-class object containing facets with 'analyzes' slot populated. Defaults to NULL}

\item{sampledata}{A data frame with samples as rows and environmental variables as columns. Needs to be unsplit. Defaults to NULL.}

\item{featurename}{A string indicating which feature is analyzed. Current possibilities: 'ASVs', 'OTUs', 'KOs', 'pathways'. Defaults to NULL.}

\item{file.name}{A string being either a filename or a common part of filenames (for beta diversity). Defaults to ''.}

\item{d}{A vector of numbers between 0 and 1 indicating which OTU dissimilarities should be processed. Needs to be given only if featurename='OTUs'. Defaults to NULL.}

\item{analysis}{A string indicating which type of analysis should be processed. Current possibilities: 'alpha-diversity', 'beta-diversity', 'diff-features'. Defaults to NULL.}

\item{variables}{A vector of strings indicating facetting variable(s). There should be one or two variables. Defaults to NULL.}

\item{vertical}{A logical indicating if produced plot(s) should be organized vertically (i.e. height should be greater than width). Defaults to TRUE.}

\item{collect}{A logical indicating if identical legends should be collected in one place. Defaults to F.}

\item{tag_level}{A string indicating how subplots should be tagged. Current possibilities: 'A'}

\item{width}{A number indicating width of a single subplot (in inches). Defaults to 3.5.}

\item{height}{A number indicating height of a single subplot (in inches). Defaults to 3.5.}
}
\value{
A ggplot2 object containing composed plot or a list containing composed plots for each metric (in case of beta diversity analysis).
}
\description{
Arrange individual plots in a grid and annotate them.
}
\examples{
NULL

}
