% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructOTUs.R
\name{constructOTUs}
\alias{constructOTUs}
\title{constructOTUs}
\usage{
constructOTUs(
  experiment = NULL,
  seqtab = NULL,
  taxonomy = NULL,
  fasta.file = "seqs.fasta",
  count.file = "seqs.count_table",
  sampledata = NULL,
  dist.file = NULL,
  basename = "",
  diss = c("0.03"),
  aligned = F,
  reference.alignment = NULL,
  start = NULL,
  stop = NULL,
  pairwise.alignment = FALSE,
  alignment.algorithm = "muscle",
  classification.db = NULL,
  generate.tree = TRUE,
  tree.algorithm = "fasttree",
  mothur = "mothur",
  fasttree = "FastTreeMP",
  raxml = "RAxML",
  muscle = "muscle",
  mafft = "mafft",
  processors = 1
)
}
\arguments{
\item{seqtab}{A named integer matrix with samples in rows and ASVs in columns. Column names should be actual ASVs' sequences.}

\item{fasta.file}{A path to a fasta file containing ASVs' sequences. Needs to be given if there is no seqtab. Defaults to 'seqs.fasta'.}

\item{count.file}{A path to a count_table file with abundances of ASVs. Needs to be given if there is no seqtab. Defaults to 'seqs.count_table'.}

\item{dist.file}{A path to a distance matrix file in column or phylip format. Optional. Defaults to NULL.}

\item{diss}{A vector of numbers between 0 and 1 indicating at which dissimilarities OTUs should be constructed. Defaults to c(0.03).}

\item{aligned}{A logical indicating if sequences in fasta.file are aligned. Defaults to FALSE.}

\item{reference.alignment}{A path to a reference alignment in fasta format (e.g. SILVA or Greengenes). Defaults to NULL.}

\item{start}{An integer indicating beginning of a kept region of an alignment or NULL in which case it will be calculated so that min. 75\% of sequences are kept. Defaults to NULL.}

\item{stop}{An integer indicating end of a kept region of an alignment or NULL in which case it will be calculated so that min. 75\% of sequences are kept. Defaults to NULL.}

\item{pairwise.alignment}{A logical indicating if pairwise alignment should be performed. Defaults to FALSE.}

\item{alignment.algorithm}{A string indicating which algorithm should be used for alignment in case no reference alignment is given, and no pairwise alignment is to be performed. One of 'muscle' or 'mafft'. Defaults to 'muscle'.}

\item{classification.db}{A path to a classification database suitable for use with dada2's assignTaxoonmy. Defaults to NULL.}

\item{generate.tree}{A logical indicating if a tree should be constructed for representative sequences. Defaults to TRUE.}

\item{tree.algorithm}{A string indicating an algorithm for tree construction. One of 'clearcut' (relaxed neighbor joining, Shenneman et al. 2006), 'fasttree' (ML by FastTreeMP) or 'raxml' (ML by RAxML). Defaults to 'fasttree'.}

\item{mothur}{A path to a MOTHUR executable. Defaults to 'mothur'.}

\item{fasttree}{A path to FastTreeMP executable. Defaults to 'FastTreeMP'.}

\item{raxml}{A path to RAxML executable. Defaults to 'RAxML'.}

\item{muscle}{A path to MUSCLE executable. Defaults to 'muscle'.}

\item{mafft}{A path to MAFFT executable. Defaults to 'mafft'.}

\item{processors}{An integer indicating how many processors should be used. Defaults to 1.}
}
\value{
A list of featuresets, one for each of dissimilarities.
}
\description{
Cluster ASVs into OTUs. Opticlust algorithm (Schloss and Westcott, 2016?) is used. Tree is computed and classification of represetative sequences is performed if needed.
}
