% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplexFastQ.R
\name{demultiplexFastQ}
\alias{demultiplexFastQ}
\title{demultiplexFastQ}
\usage{
demultiplexFastQ(
  reads1,
  reads2,
  primerF,
  primerR,
  design,
  outdir = ".",
  max_mismatch = 1,
  processors = 1
)
}
\arguments{
\item{reads1}{A path to a fastq file containig forward (R1) reads.}

\item{reads2}{A path to a fastq file containing reverse (R2) reads}

\item{primerF}{A string - forward primer sequence (the one used in the first round of PCR)}

\item{primerR}{A string - reverse primer sequence (the one used in the first round of PCR)}

\item{design}{A tsv file with three columns: samplename, index1 (the one at i5 end), index2 (the one at i7 end)}

\item{outdir}{A path to a directory where resulting files will be stored. Defaults to '.'.}

\item{max_mismatch}{An integer indicating how many mismatches are acceptable in index sequences. Defaults to 1.}

\item{processors}{Number of processors to use. Defaults to 1.}
}
\value{
Data frame with three columns: R1 file name, R2 file name, number of reads. The table contains as many rows as the number of found index combinations (if no reads were found for a combination it is not inculded in the table).
}
\description{
Demultiplex fastq paired fastq files sequenced by Novogene - pools tagged with Novogene's MIDs and within pools 'inner' MIDs. One pool at a time i.e. R1 and R2 files.
}
