% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoiseFastq.R
\name{denoiseFastQSet}
\alias{denoiseFastQSet}
\title{denoiseFastq}
\usage{
denoiseFastQSet(
  directory = ".",
  novaseqLoess = NULL,
  save = TRUE,
  basename = "bacteria",
  pattern = "_R1_001.fastq.gz",
  filtered = F,
  error.rates = NULL,
  sample.name.sep = "_",
  truncLen = "auto",
  pool = F,
  maxEE = c(2, 2),
  maxN = 0,
  quality.threshold = 25,
  max.percent.dropped = 20,
  nbases = 1e+08,
  min.length = 250,
  amplicon.length = 450,
  min.overlap = 50,
  use.mean.quality = F,
  reads = "both",
  prepareSRA = F,
  SRA.template = NULL,
  BioSample.template = NULL,
  BioProject.accession = "",
  instrument.name = "NovaSeq",
  kit.version = "1.5",
  multithread = T,
  huge = F
)
}
\arguments{
\item{directory}{A path to a directory where fastq files are stored. Defaults to '.'.}

\item{novaseqLoess}{A function to use while denoising NovaSeq data if dada2 version is below 1.32. Current possibilities: loessErrfun_mod1, loessErrfun_mod2, loessErrfun_mod3, loessErrfun_mod4. Defaults to NULL (standard version works well with NovaSeq data as of dada2 v. 1.32).}

\item{save}{A logical indicating if resulting sequence table should be saved as an RDS file. Defaults to TRUE.}

\item{basename}{A string with which names of all generated files will begin. Defaults to 'bacteria'.}

\item{pattern}{A pattern common for all fastq files to be analyzed. Defaults to '_R1_001.fastq.gz'.}

\item{filtered}{A logical indicating if precomputed filtering results should be used. If TRUE there needs to be a 'filtered' directory in the dir where fastq files are stored and it must contain filtered fastq files. Defaults to FALSE.}

\item{error.rates}{A path to an RDS file containing a list consisting of errF and errR error matrices from dada2's learnErrors(). If not NULL precomputed error rates will be used.  Defaults to NULL.}

\item{sample.name.sep}{A string separating sample name from the rest of filename. Defaults to '_'.}

\item{truncLen}{A vector of two positive integers indicating where R1 and R2 should be truncated or 'auto'. If 'auto' optimal trimming positions are calculated (see find_best_trim_lengths()). For NovaSeq usually there is no need to truncate reads, so c(0,0) or c(250,250) should be used. Defaults to 'auto'.}

\item{pool}{A logical indicating if samples should be pooled which improves rare variants detection or 'pseudo'. Increases both memory consumption and CPU time, use with huge data is discouraged. Defaults to FALSE.}

\item{quality.threshold}{A positive integer indicating quality threshold for finding best trim lengths. Needed only for Miseq. Defaults to 25.}

\item{max.percent.dropped}{A positive integer indicating maximal percentage of dropped reads for finding best trim length algorithm. Needed only for MiSeq. Defaults to 20.}

\item{nbases}{A positive integer indicating how many bases should be used in error learning process. Defaults to 10e8.}

\item{min.length}{A positive integer. Minimal length of single-end reads. Defaults to 250.}

\item{amplicon.length}{A positive integer. Length of an analyzed amplicon. Defaults to 450 (V3-V4 16S rRNA).}

\item{min.overlap}{A positive integer. Minimal length of overlap allowing reads to be merged. Defaults to 50.}

\item{use.mean.quality}{A logical indicating if mean quality should be used in finding best trim lengths algorithm. Defaults to FALSE.}

\item{reads}{A string indicating which reads should be processed. One of 'both', 'forward' (R1), 'reverse' (R2). Defaults to 'both'.}

\item{prepareSRA}{A logical indicating if table for SRA submission should be prepared. Defaults to FALSE.}

\item{SRA.template}{A path to a TSV template file. If NULL the one in package is used. Defaults to NULL.}

\item{BioSample.template}{A path to a TSV file with BioSample accessions. Defaults to NULL.}

\item{BioProject.accession}{A string indicating BioProject accession of a project from which processed samples come. Needed only if prepareSRA=TRUE. Defaults to ''.}

\item{instrument.name}{A string indicating which sequencing machine was used to generate analyzed reads. Defaults to 'NovaSeq'.}

\item{kit.version}{A string indicating which version of Illumina's sequencing kit was used to produce analyzed reads. Defaults to 'v1.5'.}

\item{multithread}{A logical indicating if multiple processors should be used or a positive integer (a number of CPUs to use). Defaults to TRUE (all processors used).}

\item{huge}{A logical indicating that samples should be processed separately. Usually needed when processing a full NovaSeq run. Defaults to TRUE.}
}
\value{
List with two elements: seqtab and report.
}
\description{
Filter, trim, denoise, and merge reads, identify and remove chimeras and construct sequence table.
}
