% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sra_download.R
\name{download_fastq_for_bioproject}
\alias{download_fastq_for_bioproject}
\title{Title}
\usage{
download_fastq_for_bioproject(bioproject_id, output_dir = ".")
}
\arguments{
\item{output_dir}{}
}
\description{
Title
}
