% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareDataFromDada2.R
\name{prepareDataFromDada2}
\alias{prepareDataFromDada2}
\title{prepareDataFromDada2}
\usage{
prepareDataFromDada2(
  seqtab = seqtab.nochim,
  taxa = NULL,
  tree = NULL,
  classification.db = NULL,
  design = NULL,
  design.sep = "\\t",
  design.rowname = NULL,
  remove.rare = 2,
  remove.taxa = c("Chloroplast", "Mitochondria"),
  keep.only = "Bacteria",
  generate.tree = TRUE,
  alignment.algorithm = "muscle",
  tree.algorithm = "fasttree",
  pairwise.alignment = FALSE,
  reference.alignment = "",
  start = NULL,
  stop = NULL,
  rarefy = 100,
  depth = "auto",
  sample.weight = 3000,
  min.depth = 500,
  step = 10,
  max.samples.discarded = 10,
  basename = "bacteria",
  rand.seed = 667,
  fasta.file = "seqs.fasta",
  count.file = "seqs.count_table",
  OTU = FALSE,
  generate.picrust = FALSE,
  picrust.filename.base = "picrust",
  conda = "conda",
  condaenv = "picrust2",
  picrust = "picrust2_pipeline.py",
  picrust.outdir = "picrust2_out",
  plot.nsti = FALSE,
  plot.nsti.var = NULL,
  marker.seq = "16S",
  save = TRUE,
  mothur = "mothur",
  mafft = "mafft",
  muscle = "muscle",
  fasttree = "FastTreeMP",
  raxml = "raxml",
  description = "",
  processors = 1,
  directory = ".",
  ...
)
}
\arguments{
\item{seqtab}{A named integer matrix with samples in rows and ASVs in columns. Column names should be actual ASVs' sequences. Chimeras should be removed prior to use of prepareDataFromDada2. Such a matrix can be produced with dada2. Defaults to seqtab.nochim.}

\item{taxa}{A data frame with ASVs in rows and taxlevels i columns or a path to an RDS file containing such an object. Can be produced with dada2. Defaults to NULL.}

\item{tree}{A phylo-class object or a path to an RDS or nexus file containing one. Defaults to NULL.}

\item{classification.db}{A path to database allowing classification with dada2's assignTaxonomy(). Defaults to NULL.}

\item{design}{A data frame with samples as rows and variables as columns or a path to a file containing such table.}

\item{design.sep}{A string indicating separator needed to read a design file. Defaults to '\\t'.}

\item{design.rowname}{A string indicating which column in design should be used as row names. Defaults to NULL.}

\item{remove.rare}{An integer indicating minimal acceptable abundance of kept ASVs. Defaults to 2 (i.e. singletons and doubletons are dropped). When working with high coverage data (e.g. off of a NovaSeq) it should be increased. At 100 000 reads/sample 5 is ok, and for 1 M seqs/sample 10 should be used.}

\item{remove.taxa}{A vector of strings with contaminating taxa names. Defaults to c('Chloroplast', 'Mitochondria').}

\item{keep.only}{A string indicating which taxonomic group should be kept. Defaults to 'Bacteria'.}

\item{generate.tree}{A logical indicating if a phylogentic tree should be computed. Defaults to TRUE.}

\item{alignment.algorithm}{A string indicating how an alignment should be computed. Current possibilities: 'mafft', 'mothur', 'muscle'. If a reference alignment is used it should be set to 'mothur'. Defaults to 'mothur'.}

\item{tree.algorithm}{A string idicating how a tree should be computed. Current possibilities: 'clearcut', 'fasttree', 'raxml'. Defaults to 'fasttree'.}

\item{pairwise.alignment}{A logical indicating if pairwise alignment should be used for tree generation. Doesn't make sense with reference alignment. Defaults to FALSE.}

\item{reference.alignment}{A path to a reference alignment, such as SILVA. MOTHUR must be used to compute alignment. Defaults to ''.}

\item{start}{NULL or an integer indicating beginning of an alignment region to keep. Makes sense only for alignments to a reference. If NULL the value is determined so that at least 90\% of sequences are kept. Defaults to NULL.}

\item{stop}{NULL or an integer indicating end of an alignment region to keep. Makes sense only for alignments to a reference. If NULL the value is determined so that at least 90\% of sequences are kept. Defaults to NULL.}

\item{rarefy}{Number of rarefaction iterations to average over. If 0 no rarefying is performed. Defaults to 100.}

\item{depth}{'auto' or an integer indicating rarefaction depth (i.e. how many randomly picked sequences from each sample to keep). If 'auto' depth is calculated so that information content is maximized. Defaults to 'auto'.}

\item{sample.weight}{An integer indicating how many sequences is 'worth' one sample, in other words, how much the algorithm should try to keep samples at the cost of decreasing depth. Defaults to 3000.}

\item{min.depth}{An integer indicating minimal acceptable rarefaction depth. Defaults to 500.}

\item{step}{An integer indicating interval at which information content will be assessed when optimizing rarefaction depth. Defaults to 10.}

\item{max.samples.discarded}{A integer indicating maximal percent of samples eliminated due to too low depth. Defaults to 10 (meaning that at least 90\% of samples must be kept).}

\item{basename}{A string being the common part of file names. Defaults to ''.}

\item{rand.seed}{An odd integer being a random numbers generator seed. Defaults to 667 (a new better beast).}

\item{fasta.file}{A path to a fasta file with ASVs sequences. If NULL the file will be generated. Defaults to NULL.}

\item{count.file}{A path to a count_table file with abundances of ASVs. If NULL the file will be generated. Defaults to NULL.}

\item{OTU}{A logical or a vector of numbers greater than 0 and lower than 1. If TRUE OTUs will be generated for 0.03 dissimilarity, if a vector of numbers is given OTUs will be computed for dissimilarities in this vector. Defaults to FALSE.}

\item{generate.picrust}{A logical indicating if PICRUSt2 analysis should be performed. Defaults to TRUE.}

\item{picrust.filename.base}{A string which will be a common part of names of PICRUSt2-produced files. Defaults to 'picrust'.}

\item{conda}{A path to conda. Defaults to 'conda'.}

\item{condaenv}{A string indicating name of a conda environment in which PICRUSt2 is installed. Defaults to 'picrust'.}

\item{picrust}{A path to PICRUSt2 script. Defaults to 'picrust2\_pipeline.py'.}

\item{picrust.outdir}{A path to a directory in which PICRUSt2 results will be stored.}

\item{plot.nsti}{A logical indicating if a plot of NSTI should be generated.}

\item{plot.nsti.var}{A string indicating which variable should be used to plot NSTI. Defaults to NULL.}

\item{marker.seq}{A string indicating what kind of marker sequence the experiment uses. Defaults to '16S'.}

\item{save}{A logical indicating of the resulting object should be saved as an RDS file.}

\item{mothur}{A path to MOTHUR executable. Defaults to 'mothur'.}

\item{mafft}{A path to MAFFT executable. Defaults to 'mafft'.}

\item{muscle}{A path to MUSCLE executable. Defaults to 'muscle'.}

\item{fasttree}{A path to FastTree executable. Defaults to 'FastTreeMP'.}

\item{raxml}{A path to RAxML executable. Defaults to 'raxml'.}

\item{description}{A string with description of the analyzed experiment.}

\item{processors}{An integer indicating number of processors to use. Defaults to 1.}

\item{...}{Other arguments passed on to worker functions.}
}
\value{
An mg-class object.
}
\description{
Prepare an mg-class object based on dada2 results.
}
