% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitFastQNovogene.R
\name{splitFastQNovogene}
\alias{splitFastQNovogene}
\title{Split paired fastq files according to smaple indexes}
\usage{
splitFastQNovogene(
  reads1,
  reads2,
  primerF,
  primerR,
  outfile_R1,
  outfile_R2,
  ind_f,
  ind_r,
  max_mismatch
)
}
\arguments{
\item{max_mismatch}{}
}
\description{
Split paired fastq files according to smaple indexes
}
